Update CPP Module 07 : ex02

+++ CPP Module 07/ex02/Array.hpp
@@ -6,7 +6,7 @@ template<typename T>
 class Array{
 private:
 	T *_array;
-	unsigned int _size;
+	size_t _size;
 
 public:
 	Array(): _size(0){
@@ -14,7 +14,7 @@ public:
 		this->_array = new T[this->_size];
 	}
 
-	Array(unsigned int size): _size(size){
+	Array(size_t size): _size(size){
 		std::cout << "Constructor for an Array of size " << size << " called" << std::endl;
 		this->_array = new T[this->_size];
 	}
@@ -36,13 +36,13 @@ public:
 		if(src.size() != 0){
 			this->_size = src.size();
 			this->_array = new T[this->_size];
-			for(unsigned int i = 0; i < this->size(); i++)
+			for(size_t i = 0; i < this->size(); i++)
 				this->_array[i] = src._array[i];
 		}
 		return (*this);
 	}
 
-	T& operator[](unsigned int index){
+	T& operator[](size_t index){
 		if(index >= this->_size || this->_array == NULL){
 			std::cout << "index: " << index << std::endl;
 			throw Array<T>::OutOfIndexException();
@@ -55,7 +55,7 @@ public:
 		virtual const char	*what() const throw();
 	};
 
-	unsigned int size() const{
+	size_t size() const{
 		return (this->_size);
 	}
 };
